+++ optd-dsl/src/analyzer/mod.rs
@@ -2,6 +2,6 @@ pub mod context;
 pub mod from_ast;
 pub mod hir;
 pub mod map;
-pub mod semantic_checker;
-pub mod type_checker;
+pub mod semantic_check;
+pub mod type_check;
 pub mod types;
