--- optd-dsl/src/analyzer/semantic_checker/mod.rs
@@ -1,1 +1,0 @@
-pub mod error;
