--- optd-dsl/src/analyzer/type_checker/mod.rs
@@ -1,1 +1,0 @@
-pub mod error;
